Consolidate shared stratum socket, wifi and clean-queue helpers
Move the duplicated socket resolve/options code into a new
components/stratum/stratum_socket module, is_wifi_connected into the
connect component (wifi_is_connected), and the clean-jobs-queue logic
into system.c (SYSTEM_clean_jobs_queue). SV1 and SV2 now share these.

SV2 also resolves the pool address up front and connects by IP so DNS
stays non-blocking, matching SV1.
