+++ components/connect/include/connect.h
@@ -4,6 +4,7 @@
 #include "lwip/sys.h"
 #include <arpa/inet.h>
 #include <lwip/netdb.h>
+#include <stdbool.h>
 
 #include "esp_err.h"
 #include "esp_wifi_types.h"
@@ -20,5 +21,6 @@ void wifi_init(void * GLOBAL_STATE);
 esp_err_t wifi_apply_hostname(const char *hostname);
 esp_err_t wifi_scan(wifi_ap_record_simple_t *ap_records, uint16_t *ap_count);
 esp_err_t get_wifi_current_rssi(int8_t *rssi);
+bool wifi_is_connected(void);
 
 #endif /* CONNECT_H_ */
