+++ components/stratum/CMakeLists.txt
@@ -3,10 +3,11 @@ SRCS
     "utils.c"
     "mining.c"
     "stratum_api.c"
+    "stratum_socket.c"
     "coinbase_decoder.c"
     "segwit_addr.c"
     "base58.c"
-                    
+
 INCLUDE_DIRS
     "include"
 
@@ -16,4 +17,5 @@ REQUIRES
     "app_update"
     "esp_timer"
     "tcp_transport"
+    "esp_netif"
 )
