+++ main/system.h
@@ -8,6 +8,11 @@ void SYSTEM_init_system(GlobalState * GLOBAL_STATE);
 void SYSTEM_init_versions(GlobalState * GLOBAL_STATE);
 esp_err_t SYSTEM_init_peripherals(GlobalState * GLOBAL_STATE);
 
+// Clear the stratum job queue and valid-job tracking on a clean-jobs event,
+// and reset hashrate measurements so reconnects don't spike the average.
+// Shared by the SV1 and SV2 tasks.
+void SYSTEM_clean_jobs_queue(GlobalState * GLOBAL_STATE);
+
 void SYSTEM_notify_accepted_share(GlobalState * GLOBAL_STATE);
 void SYSTEM_notify_rejected_share(GlobalState * GLOBAL_STATE, char * error_msg);
 void SYSTEM_notify_found_nonce(GlobalState * GLOBAL_STATE, double diff, uint8_t job_id);
